move authOptions into its own file
